Refactor backend package.
