+++ spacy_llm/backends/integration/__init__.py
@@ -1,8 +1,9 @@
-from .base import RemoteBackend, HuggingFaceBackend
+from .base import RemoteBackend
 from .langchain import backend_langchain, query_langchain
 from .minichain import backend_minichain, query_minichain
-from .dolly import backend_dolly_hf
-from .stablelm import backend_stablelm_hf
+from .hf.base import HuggingFaceBackend
+from .hf.dolly import backend_dolly_hf
+from .hf.stablelm import backend_stablelm_hf
 
 __all__ = [
     "RemoteBackend",
