+++ spacy_llm/backends/integration/hf/__init__.py
@@ -1,0 +1,9 @@
+from .base import HuggingFaceBackend
+from .dolly import backend_dolly_hf
+from .stablelm import backend_stablelm_hf
+
+__all__ = [
+    "HuggingFaceBackend",
+    "backend_dolly_hf",
+    "backend_stablelm_hf",
+]
