+++ spacy_llm/backends/rest/__init__.py
@@ -1,3 +1,4 @@
-from .registry import backend_rest
+from . import base, openai, noop
+from .registry import backend_rest, supported_apis
 
-__all__ = ["backend_rest"]
+__all__ = ["base", "openai", "noop", "supported_apis", "backend_rest"]
